content-tile: Protect against contents without an artist
Some contents (say MexChannel) don't provide a property name for
artists. Let's test for NULL before trying to bind it.

Took the opportunity to fix the mixup between the names of the variables
holding the property names. TITLE is bound to the primary label while
ARTIST is bound to the secondary label.