channel-manager: Move the "live" category to ChannelManager
ChannelManager is the class responsible for the list of channels. It's
only natural to move the "live" category there. For instance the mapping
between the channel number and the channel will be done there (and it'll
be one way to sort channels).

This breaks the gnome-dvb plugin that relies on the "live" categorty to
be created but does not use ChannelManager.